requirements.txt
pyyaml==5.4.1
requests==2.25.1
SPARQLWrapper==1.8.5
pandas==1.2.4
jupyter==1.0.0
bokeh==2.3.2
matplotlib==3.4.2
scipy
wikipedia==1.4.0
spacy==3.0.6  # python -m spacy download en_core_web_sm
scikit-learn==0.24.2
nltk==3.6.2  # import nltk; nltk.download("stopwords")
wordcloud==1.8.1
gensim==4.0.1
python-Levenshtein==0.12.2
lxml==4.6.3
html5lib==1.1
wptools==0.4.17
streamlit==0.87.0
pyvis==0.1.9
streamlit-agraph==0.0.35
seaborn==0.11.2
rdflib==5.0.0
streamlit-pipeline==0.0.2
PrettyTable==3.2.0
